Refactored file names of componenets to fit typical react casing (#461)
* refactor: moved file to have capital D

* refactor: moved the files to different names that match the common casing for components

* nit: updated portal path aswell in index.tsx

* fix: portal was being imported here too
